Rename rfnm_api.h to rfnm_fw_api.h
To avoid confusion
