+++ include/librfnm/device.h
@@ -8,7 +8,7 @@
 #include <array>
 #include <vector>
 
-#include "rfnm_api.h"
+#include "rfnm_fw_api.h"
 
 #if defined(__GNUC__)
 #define RFNM_PACKED_STRUCT( __Declaration__ ) __Declaration__ __attribute__((__packed__))
